+++ packages/traffic-sign-converter/src/trafficSignTagToSigns/trafficSignTagToSigns.ts
@@ -90,7 +90,7 @@ export const trafficSignTagToSigns = (
       signId: null,
       descriptiveName: osmValuePart,
       // When the initial string (`cleaned`) includes our unkown values when split by the `traffic_sign` separator (`;`) then it is this, all else is a `modifier_sign`; which would include `traffic_sign=foobar` as modifier.
-      kind: cleaned.split(';').includes(osmValuePart) ? 'traffic_sign' : 'modifier_sign',
+      kind: cleaned.split(';').includes(osmValuePart) ? 'traffic_sign' : 'condition_modifier',
     }
     return unkownSign
   })